requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.22
bcrypt==3.2.0
python-dotenv==0.19.0
scikit-learn==0.24.2